Add age field support for user object in token endpoint.

+++ src/main/java/com/github/pt/token/UserResponseDTO.java
@@ -16,5 +16,5 @@ public class UserResponseDTO {
     String name;
     String avatar;    
     String gender;
-    Integer age;
+    Long age;
 }
